# src/main/resources/static/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Rest API
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/v1/users/{userId}:
    get:
      tags:
        - user-controller
      operationId: getUser
      summary:
        "Return user by id (ADMIN)"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      summary:
        "Delete user by id (ADMIN)"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
    patch:
      tags:
        - user-controller
      operationId: updateUser
      summary:
        "Update user by id (ADMIN)"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
...
  "/api/v1/notifications/{id}":
    delete:
      tags:
        - notification-controller
      operationId: delete
      summary:
        "Delete notification (ADMIN)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            "*/*":
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                type: string
      security:
        - bearerAuth: [ ]
  "/api/v1/notifications/users/{userId}":
    get:
      tags:
        - notification-controller
      operationId: getByUser
      summary:
        "Get notifications by user_id (USER)"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/NotificationResponse"
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            "*/*":
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                type: string

      security:
        - bearerAuth: [ ]
  "/api/v1/notifications":
    post:
      tags:
        - notification-controller
      operationId: create
      summary:
        "Create notification (ADMIN)"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NotificationRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/NotificationResponse"
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            "*/*":
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                type: string
      security:
        - bearerAuth: [ ]
  "/api/v1/notifications/{id}/read":
    put:
      tags:
        - notification-controller
      operationId: markRead
      summary:
        "Mark notification as read (USER)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/NotificationResponse"
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            "*/*":
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                type: string

      security:
        - bearerAuth: [ ]
components:
  schemas:
    NotificationRequest:
      type: object
      properties:
        type:
          type: string
          enum:
            - INFO
            - SUCCESS
            - WARNING
        message:
          type: string
          maxLength: 2147483647
          minLength: 2
        userId:
          type: integer
          format: int64
      required:
        - type
        - userId
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - INFO
            - SUCCESS
            - WARNING
        message:
          type: string
        read:
          type: boolean
        timestamp:
          type: string
          format: date-time
        userId:
          type: integer
          format: int64
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 30
          minLength: 2
        surname:
          type: string
          maxLength: 30
          minLength: 2
        email:
          type: string
...
    UserUpdateDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 30
          minLength: 2
        surname:
          type: string
          maxLength: 30
          minLength: 2
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer